refactor(token-provisioning): own creation model types
