+++ src/services/apiService/newApiFamily/keyManagement.ts
@@ -1,13 +1,13 @@
 import { SITE_TYPES, type AccountSiteType } from "~/constants/siteType"
 import * as commonKeyManagement from "~/services/apiService/common"
+import type { ApiServiceRequest } from "~/services/apiService/common/type"
+import * as oneHubKeyManagement from "~/services/apiService/oneHub"
+import * as wongKeyManagement from "~/services/apiService/wong"
 import type {
-  ApiServiceRequest,
   CreateTokenRequest,
   CreateTokenResult,
   UserGroupInfo,
-} from "~/services/apiService/common/type"
-import * as oneHubKeyManagement from "~/services/apiService/oneHub"
-import * as wongKeyManagement from "~/services/apiService/wong"
+} from "~/services/tokenProvisioning/model"
 import type { ApiToken } from "~/types"
 
 interface KeyManagementImplementation {
