+++ src/services/apiService/oneHub/index.ts
@@ -1,8 +1,5 @@
 import { normalizeApiTokenKey } from "~/services/accountTokens/apiTokenKey"
-import {
-  type ApiServiceRequest,
-  type UserGroupInfo,
-} from "~/services/apiService/common/type"
+import type { ApiServiceRequest } from "~/services/apiService/common/type"
 import { fetchApiData } from "~/services/apiService/common/utils"
 import {
   transformModelPricing,
@@ -15,6 +12,7 @@ import type {
   PaginatedTokenDate,
 } from "~/services/apiService/oneHub/type"
 import type { PricingResponse } from "~/services/modelList/pricingModel"
+import type { UserGroupInfo } from "~/services/tokenProvisioning/model"
 import type { ApiToken } from "~/types"
 import { createLogger } from "~/utils/core/logger"
