+++ tests/services/apiService/aihubmix/index.test.ts
@@ -28,9 +28,9 @@ import {
   validateAccountConnection,
 } from "~/services/apiService/aihubmix"
 import { API_ERROR_CODES, ApiError } from "~/services/apiService/common/errors"
-import { type CreateTokenRequest } from "~/services/apiService/common/type"
 import { MODEL_LIST_SOURCE_KINDS } from "~/services/modelList/pricingModel"
 import { calculateModelPrice } from "~/services/models/utils/modelPricing"
+import type { CreateTokenRequest } from "~/services/tokenProvisioning/model"
 import { AuthTypeEnum } from "~/types"
 import { server } from "~~/tests/msw/server"
