+++ tests/services/apiService/sub2api/keyManagement.test.ts
@@ -2,10 +2,7 @@ import { beforeEach, describe, expect, it, vi } from "vitest"
 
 import { ACCOUNT_BROWSER_SESSION_SOURCES } from "~/services/accountBrowserSession/types"
 import { ApiError } from "~/services/apiService/common/errors"
-import type {
-  ApiServiceRequest,
-  CreateTokenRequest,
-} from "~/services/apiService/common/type"
+import type { ApiServiceRequest } from "~/services/apiService/common/type"
 import {
   createApiToken,
   fetchAccountAvailableModels,
@@ -23,6 +20,7 @@ import {
   translateSub2ApiCreateTokenRequest,
   translateSub2ApiUpdateTokenRequest,
 } from "~/services/apiService/sub2api/parsing"
+import type { CreateTokenRequest } from "~/services/tokenProvisioning/model"
 import { AuthTypeEnum } from "~/types"
 
 const {
